+++ pmd-compat6/src/main/java/net/sourceforge/pmd/RuleSetFactory.java
@@ -2,6 +2,9 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
+// This class has been taken from 7.0.0-SNAPSHOT
+// Before removing RuleSetFactoryCompatibility (#4314)
+
 package net.sourceforge.pmd;
 
 import static net.sourceforge.pmd.util.CollectionUtil.setOf;
@@ -80,6 +83,7 @@ final class RuleSetFactory {
     private final LanguageRegistry languageRegistry;
     private final RulePriority minimumPriority;
     private final boolean warnDeprecated;
+    private final RuleSetFactoryCompatibility compatibilityFilter;
     private final MessageReporter reporter;
     private final boolean includeDeprecatedRuleReferences;
 
@@ -89,6 +93,7 @@ final class RuleSetFactory {
                    LanguageRegistry languageRegistry,
                    RulePriority minimumPriority,
                    boolean warnDeprecated,
+                   RuleSetFactoryCompatibility compatFilter,
                    boolean includeDeprecatedRuleReferences,
                    MessageReporter reporter) {
         this.resourceLoader = resourceLoader;
@@ -97,6 +102,7 @@ final class RuleSetFactory {
         this.warnDeprecated = warnDeprecated;
         this.includeDeprecatedRuleReferences = includeDeprecatedRuleReferences;
 
+        this.compatibilityFilter = compatFilter;
         this.reporter = reporter;
     }
 
@@ -162,9 +168,9 @@ final class RuleSetFactory {
     private @NonNull RuleSet readDocument(RuleSetReferenceId ruleSetReferenceId, boolean withDeprecatedRuleReferences) {
 
         try (CheckedInputStream inputStream = new CheckedInputStream(ruleSetReferenceId.getInputStream(resourceLoader), new Adler32())) {
-            if (!ruleSetReferenceId.isAbsolute()) {
+            if (!ruleSetReferenceId.isExternal()) {
                 throw new IllegalArgumentException(
-                    "Cannot parse a RuleSet from a non-absolute reference: <" + ruleSetReferenceId + ">.");
+                    "Cannot parse a RuleSet from a non-external reference: <" + ruleSetReferenceId + ">.");
             }
 
             XmlMessageHandler printer = getXmlMessagePrinter();
@@ -371,6 +377,7 @@ final class RuleSetFactory {
                     // has been reported
                     continue;
                 }
+                excludedRuleName = compatibilityFilter.applyExclude(ref, excludedRuleName, this.warnDeprecated);
                 if (excludedRuleName != null) {
                     excludedRulesCheck.put(excludedRuleName, child);
                 }
@@ -437,12 +444,15 @@ final class RuleSetFactory {
     private RuleSetReferenceId parseReferenceAndWarn(String ref,
                                                      Node xmlPlace,
                                                      PmdXmlReporter err) {
+        ref = compatibilityFilter.applyRef(ref, this.warnDeprecated);
         if (ref == null) {
             err.at(xmlPlace).warn("Rule reference references a deleted rule, ignoring");
             return null; // deleted rule
         }
+        // only emit a warning if we check for deprecated syntax
+        MessageReporter subReporter = warnDeprecated ? err.at(xmlPlace) : MessageReporter.quiet();
 
-        List<RuleSetReferenceId> references = RuleSetReferenceId.parse(ref);
+        List<RuleSetReferenceId> references = RuleSetReferenceId.parse(ref, subReporter);
         if (references.size() > 1 && warnDeprecated) {
             err.at(xmlPlace).warn("Using a comma separated list as a ref attribute is deprecated. "
                                       + "All references but the first are ignored.");
@@ -520,13 +530,13 @@ final class RuleSetFactory {
         RuleSetFactory ruleSetFactory = toLoader().filterAbovePriority(RulePriority.LOW).warnDeprecated(false).toFactory();
 
         boolean isSameRuleSet = false;
-        if (!otherRuleSetReferenceId.isAbsolute()
+        if (!otherRuleSetReferenceId.isExternal()
             && containsRule(ruleSetReferenceId, otherRuleSetReferenceId.getRuleName())) {
-            otherRuleSetReferenceId = new RuleSetReferenceId(ref, ruleSetReferenceId);
+            otherRuleSetReferenceId = new RuleSetReferenceId(ref, ruleSetReferenceId, err.at(REF.getAttributeNode(ruleNode)));
             isSameRuleSet = true;
-        } else if (otherRuleSetReferenceId.isAbsolute()
+        } else if (otherRuleSetReferenceId.isExternal()
             && otherRuleSetReferenceId.getRuleSetFileName().equals(ruleSetReferenceId.getRuleSetFileName())) {
-            otherRuleSetReferenceId = new RuleSetReferenceId(otherRuleSetReferenceId.getRuleName(), ruleSetReferenceId);
+            otherRuleSetReferenceId = new RuleSetReferenceId(otherRuleSetReferenceId.getRuleName(), ruleSetReferenceId, err.at(REF.getAttributeNode(ruleNode)));
             isSameRuleSet = true;
         }
         // do not ignore deprecated rule references
@@ -657,6 +667,7 @@ final class RuleSetFactory {
         return new RuleSetLoader().loadResourcesWith(resourceLoader)
                                   .filterAbovePriority(minimumPriority)
                                   .warnDeprecated(warnDeprecated)
+                                  .enableCompatibility(compatibilityFilter != null)
                                   .includeDeprecatedRuleReferences(includeDeprecatedRuleReferences);
     }